# specs/001-build-a-cross/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Zanari Savings & Payments API
  description: Cross-platform mobile app API for automated savings through round-ups and M-Pesa payments
  version: 1.0.0
  contact:
    name: Zanari API Team
    email: api@zanari.app
servers:
  - url: https://api.zanari.app/v1
    description: Production server
  - url: https://staging-api.zanari.app/v1  
    description: Staging server
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Authenticate user with email/phone
      description: Initial authentication step using email OTP or phone SMS
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: sarah@example.com
                phone:
                  type: string
                  pattern: '^254[0-9]{9}$'
                  example: "254712345678"
              oneOf:
                - required: [email]
                - required: [phone]
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent to your email"
                  session_id:
                    type: string
                    example: "sess_abc123"
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimit'
  /auth/verify-otp:
    post:
      summary: Verify OTP and complete authentication
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - otp_code
              properties:
                session_id:
                  type: string
                  example: "sess_abc123"
                otp_code:
                  type: string
                  pattern: '^[0-9]{6}$'
                  example: "123456"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                  refresh_token:
                    type: string
                    example: "rt_abc123def456"
                  user:
                    $ref: '#/components/schemas/User'
                  requires_pin_setup:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/setup-pin:
    post:
      summary: Set up 4-digit PIN for transactions
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pin
                - confirm_pin
              properties:
                pin:
                  type: string
                  pattern: '^[0-9]{4}$'
                  example: "1234"
                confirm_pin:
                  type: string
                  pattern: '^[0-9]{4}$'
                  example: "1234"
      responses:
        '200':
          description: PIN setup successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "PIN setup completed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/verify-pin:
    post:
      summary: Verify PIN for transaction authorization
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pin
              properties:
                pin:
                  type: string
                  pattern: '^[0-9]{4}$'
                  example: "1234"
      responses:
        '200':
          description: PIN verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                    example: true
                  token:
                    type: string
                    description: Short-lived token for transaction authorization
                    example: "txn_abc123"
        '401':
          description: Invalid PIN or account locked
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid PIN"
                  attempts_remaining:
                    type: integer
                    example: 2
                  locked_until:
                    type: string
                    format: date-time
                    example: "2025-09-23T14:30:00Z"

  # Wallet Management
  /wallets:
    get:
      summary: Get user's wallet balances
      tags:
        - Wallets
      responses:
        '200':
          description: Wallet balances retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wallet'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /wallets/{walletId}/withdraw:
    post:
      summary: Withdraw from wallet (with settlement delay for savings)
      tags:
        - Wallets
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            example: "wallet_123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - pin_token
              properties:
                amount:
                  type: number
                  minimum: 100
                  example: 50000
                  description: Amount in cents (KES)
                pin_token:
                  type: string
                  example: "txn_abc123"
                mpesa_phone:
                  type: string
                  pattern: '^254[0-9]{9}$'
                  example: "254712345678"
      responses:
        '200':
          description: Withdrawal initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id:
                    type: string
                    example: "txn_withdraw_123"
                  settlement_delay_minutes:
                    type: integer
                    example: 2
                  estimated_completion:
                    type: string
                    format: date-time
                    example: "2025-09-23T14:32:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Transaction Management
  /transactions:
    get:
      summary: Get transaction history
      tags:
        - Transactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: type
          in: query
          schema:
            type: string
            enum: [payment, transfer_in, transfer_out, round_up, bill_payment, withdrawal, deposit]
        - name: category
          in: query
          schema:
            type: string
            enum: [airtime, groceries, school_fees, utilities, transport, entertainment, savings, transfer, other]
      responses:
        '200':
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 150
                      limit:
                        type: integer
                        example: 20
                      offset:
                        type: integer
                        example: 0
                      has_more:
                        type: boolean
                        example: true
  /transactions/{transactionId}:
    get:
      summary: Get transaction details
      tags:
        - Transactions
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            example: "txn_123"
      responses:
        '200':
          description: Transaction details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'

  # Payments
  /payments/merchant:
    post:
      summary: Make payment to merchant (with automatic round-up)
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - pin_token
                - merchant_info
              properties:
                amount:
                  type: number
                  minimum: 100
                  maximum: 500000
                  example: 23000
                  description: Payment amount in cents (KES)
                pin_token:
                  type: string
                  example: "txn_abc123"
                merchant_info:
                  type: object
                  properties:
                    name:
                      type: string
                      example: "Java House"
                    till_number:
                      type: string
                      example: "123456"
                    paybill_number:
                      type: string
                      example: "400200"
                    account_number:
                      type: string
                      example: "ACC001"
                  oneOf:
                    - required: [name, till_number]
                    - required: [name, paybill_number, account_number]
                description:
                  type: string
                  example: "Lunch payment"
      responses:
        '200':
          description: Payment initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  payment_transaction_id:
                    type: string
                    example: "txn_payment_123"
                  round_up_transaction_id:
                    type: string
                    example: "txn_roundup_456"
                  total_charged:
                    type: number
                    example: 25000
                    description: Original amount + round-up
                  round_up_amount:
                    type: number
                    example: 2000
                    description: Amount saved to savings wallet
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Insufficient funds
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Insufficient funds"
                  available_balance:
                    type: number
                    example: 20000
                  round_up_skipped:
                    type: boolean
                    example: true
  /payments/transfer:
    post:
      summary: Send money to another user (P2P transfer)
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - pin_token
                - recipient
              properties:
                amount:
                  type: number
                  minimum: 100
                  maximum: 500000
                  example: 50000
                pin_token:
                  type: string
                  example: "txn_abc123"
                recipient:
                  type: object
                  properties:
                    phone:
                      type: string
                      pattern: '^254[0-9]{9}$'
                      example: "254712345678"
                    email:
                      type: string
                      format: email
                      example: "friend@example.com"
                  oneOf:
                    - required: [phone]
                    - required: [email]
                description:
                  type: string
                  example: "Money for textbooks"
      responses:
        '200':
          description: Transfer initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transfer_transaction_id:
                    type: string
                    example: "txn_transfer_123"
                  round_up_transaction_id:
                    type: string
                    example: "txn_roundup_456"
                  total_charged:
                    type: number
                    example: 51000
                  round_up_amount:
                    type: number
                    example: 1000

  # Savings Goals
  /savings-goals:
    get:
      summary: Get user's savings goals
      tags:
        - Savings Goals
      responses:
        '200':
          description: Savings goals retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  goals:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavingsGoal'

    post:
      summary: Create new savings goal
      tags:
        - Savings Goals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - target_amount
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "Emergency Fund"
                description:
                  type: string
                  maxLength: 200
                  example: "Save for unexpected expenses"
                target_amount:
                  type: number
                  minimum: 100
                  example: 500000
                  description: Target amount in cents (KES)
                target_date:
                  type: string
                  format: date
                  example: "2025-12-31"
                lock_in_enabled:
                  type: boolean
                  default: false
                  example: false
      responses:
        '201':
          description: Savings goal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'
  /savings-goals/{goalId}:
    put:
      summary: Update savings goal
      tags:
        - Savings Goals
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            example: "goal_123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                description:
                  type: string
                  maxLength: 200
                target_amount:
                  type: number
                  minimum: 100
                target_date:
                  type: string
                  format: date
                status:
                  type: string
                  enum: [active, paused, cancelled]
      responses:
        '200':
          description: Savings goal updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsGoal'

  # Round-up Rules
  /round-up-rules:
    get:
      summary: Get user's round-up configuration
      tags:
        - Round-up Rules
      responses:
        '200':
          description: Round-up rules retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundUpRule'

    put:
      summary: Update round-up configuration
      tags:
        - Round-up Rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                increment_type:
                  type: string
                  enum: ['10', '50', '100', 'auto']
                  example: "10"
                is_enabled:
                  type: boolean
                  example: true
                auto_settings:
                  type: object
                  properties:
                    min_increment:
                      type: number
                      example: 500
                    max_increment:
                      type: number
                      example: 5000
                    analysis_period_days:
                      type: integer
                      minimum: 7
                      maximum: 90
                      example: 30
      responses:
        '200':
          description: Round-up rules updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoundUpRule'

  # KYC Management
  /kyc/documents:
    get:
      summary: Get KYC document status
      tags:
        - KYC
      responses:
        '200':
          description: KYC documents retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/KYCDocument'
                  kyc_status:
                    type: string
                    enum: [not_started, pending, approved, rejected]
                    example: "pending"

    post:
      summary: Upload KYC document
      tags:
        - KYC
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - document_type
                - file
              properties:
                document_type:
                  type: string
                  enum: [national_id, passport, driving_license, selfie]
                  example: "national_id"
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: KYC document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCDocument'
        '400':
          description: Invalid file format or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        email:
          type: string
          format: email
          example: "sarah@example.com"
        phone:
          type: string
          example: "254712345678"
        first_name:
          type: string
          example: "Sarah"
        last_name:
          type: string
          example: "Mutindi"
        kyc_status:
          type: string
          enum: [not_started, pending, approved, rejected]
          example: "approved"
        status:
          type: string
          enum: [active, suspended, closed]
          example: "active"
        notification_preferences:
          type: object
          properties:
            push_enabled:
              type: boolean
              example: true
            email_enabled:
              type: boolean
              example: true
            transaction_alerts:
              type: boolean
              example: true
            savings_milestones:
              type: boolean
              example: true
        created_at:
          type: string
          format: date-time
          example: "2025-09-01T10:00:00Z"
    Wallet:
      type: object
      properties:
        id:
          type: string
          example: "wallet_main_123"
        wallet_type:
          type: string
          enum: [main, savings]
          example: "main"
        balance:
          type: number
          example: 125000
          description: Balance in cents (KES)
        available_balance:
          type: number
          example: 120000
          description: Available balance minus pending transactions
        last_transaction_at:
          type: string
          format: date-time
          example: "2025-09-23T09:30:00Z"
    Transaction:
      type: object
      properties:
        id:
          type: string
          example: "txn_123"
        type:
          type: string
          enum: [payment, transfer_in, transfer_out, round_up, bill_payment, withdrawal, deposit]
          example: "payment"
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
          example: "completed"
        amount:
          type: number
          example: 23000
          description: Amount in cents (KES)
        fee:
          type: number
          example: 0
          description: Transaction fee in cents
        description:
          type: string
          example: "Payment to Java House"
        category:
          type: string
          enum: [airtime, groceries, school_fees, utilities, transport, entertainment, savings, transfer, other]
          example: "groceries"
        merchant_info:
          type: object
          properties:
            name:
              type: string
              example: "Java House"
            till_number:
              type: string
              example: "123456"
        round_up_details:
          type: object
          properties:
            original_amount:
              type: number
              example: 23000
            round_up_amount:
              type: number
              example: 2000
            round_up_rule:
              type: string
              example: "25000"
            related_transaction_id:
              type: string
              example: "txn_roundup_456"
        created_at:
          type: string
          format: date-time
          example: "2025-09-23T09:30:00Z"
        completed_at:
          type: string
          format: date-time
          example: "2025-09-23T09:30:05Z"
    SavingsGoal:
      type: object
      properties:
        id:
          type: string
          example: "goal_123"
        name:
          type: string
          example: "Emergency Fund"
        description:
          type: string
          example: "Save for unexpected expenses"
        target_amount:
          type: number
          example: 500000
          description: Target amount in cents (KES)
        current_amount:
          type: number
          example: 125000
          description: Current progress in cents
        progress_percentage:
          type: number
          example: 25.0
          description: Progress as percentage (0-100)
        target_date:
          type: string
          format: date
          example: "2025-12-31"
        status:
          type: string
          enum: [active, completed, paused, cancelled]
          example: "active"
        lock_in_enabled:
          type: boolean
          example: false
        milestones:
          type: array
          items:
            type: object
            properties:
              percentage:
                type: number
                example: 25
              amount:
                type: number
                example: 125000
              reached_at:
                type: string
                format: date-time
                example: "2025-09-23T10:00:00Z"
              celebrated:
                type: boolean
                example: true
        created_at:
          type: string
          format: date-time
          example: "2025-09-01T10:00:00Z"
...
    KYCDocument:
      type: object
      properties:
        id:
          type: string
          example: "doc_123"
        document_type:
          type: string
          enum: [national_id, passport, driving_license, selfie]
          example: "national_id"
        status:
          type: string
          enum: [uploaded, processing, approved, rejected]
          example: "approved"
        verification_notes:
          type: string
          example: "Document verified successfully"
        uploaded_at:
          type: string
          format: date-time
          example: "2025-09-01T10:00:00Z"
        processed_at:
          type: string
          format: date-time
          example: "2025-09-01T12:00:00Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        message:
          type: string
          example: "The provided data is invalid"
        code:
          type: string
          example: "INVALID_DATA"
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimit:
      description: Rate Limit Exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...